chore(nextjs): Introduce shared module to contain @clerk/shared related code
