+++ packages/nextjs/src/shared/index.ts
@@ -1,0 +1,1 @@
+export { isHttpOrHttps, handleValueOrFn } from './url';
